fix file metadata population/serialization

+++ src/main/java/com/cloudbeats/dto/AudioFileMetadataDto.java
@@ -8,7 +8,7 @@ public record AudioFileMetadataDto(
         String album,
         List<String> genres,
         String albumCoverUrl,
-        Double duration,
+        Integer duration,
         String previewUrl
 ){
 }
